nb-puller/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
elastic-transport==8.4.0
elasticsearch==8.6.2
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
mariadb==1.1.6
multidict==6.0.4
packaging==23.0
pydantic==1.10.5
requests==2.28.2
smmap==5.0.0
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
